Refactor changes
