+++ lib/aggregation/accumulator/src/index.js
@@ -31,9 +31,9 @@ const rconfig = yieldable(rconfigcb);
 const windowsSizes = process.env.TIME_WINDOWS_SIZES ?
   JSON.parse(process.env.TIME_WINDOWS_SIZES) : undefined;
 
-const createTimeWindows = require('./time-windows');
+const createTimeWindows = timewindow.timeWindowsSizes;
 
-const dimension = createTimeWindows.dimension;
+const dimension = timewindow.dimension;
 
 // The scaling factor of each time window for creating the date string
 // [Second, Minute, Hour, Day, Month]
