+++ lib/utils/timewindow/src/index.js
@@ -6,6 +6,12 @@ const _ = require('underscore');
 const map = _.map;
 const rest = _.rest;
 
+const sec = 's';
+const min = 'm';
+const hour = 'h';
+const day = 'D';
+const month = 'M';
+
 // Setup debug log
 const debug = require('abacus-debug')('abacus-timewindow');
 
@@ -20,8 +26,43 @@ const shortcut = {
   s: 's'
 };
 
+// Time dimension keys as constants
+const timeDimension = {
+  sec,
+  min,
+  hour,
+  day,
+  month
+};
+
 // Time dimension keys corresponding to their respective window positions
-const dimensions = ['s', 'm', 'h', 'D', 'M'];
+const dimensions = [sec, min, hour, day, month];
+
+// Used to calculate time windows based on slack size
+const slackscale = {
+  [month]: { [month]: 1 },
+  [day]:   { [month]: 28, [day]: 1 },
+  [hour]:  { [month]: 672, [day]: 24, [hour]: 1 },
+  [min]:   { [month]: 40320, [day]: 1440, [hour]: 60, [min]: 1 },
+  [sec]:   { [month]: 2419200, [day]: 86400, [hour]: 3600, [min]: 60, [sec]: 1 }
+};
+
+const timeWindowsSizes = (slack, windowsSizes) => {
+  const slackBasedWindows = (dimension) => {
+    if(slack.scale && slackscale[slack.scale][dimension])
+      return map(Array(Math.ceil(
+        1 / slackscale[slack.scale][dimension] * slack.width) + 1), () => null);
+    return [null];
+  };
+
+  return {
+    getWindows: (dimension) => {
+      if (windowsSizes && windowsSizes[dimension])
+        return map(Array(windowsSizes[dimension]), () => null);
+      return slackBasedWindows(dimension);
+    }
+  };
+};
 
 // Takes a Date or moment.js object and returns a Javascript Date equivalent
 // with all dimensions lower than the given one set to 0
@@ -112,3 +153,6 @@ module.exports.timeWindowBounds = timeWindowBounds;
 module.exports.shiftWindow = shiftWindow;
 module.exports.cellfn = cellfn;
 module.exports.adjustWindows = adjustWindows;
+module.exports.dimension = timeDimension;
+module.exports.timeWindowsSizes = timeWindowsSizes;
+
