+++ lib/utils/timewindow/src/test/test.js
@@ -5,6 +5,8 @@
 
 const moment = require('abacus-moment');
 const time = require('..');
+const createTimeWindow = time.timeWindowsSizes;
+const dimension = time.dimension;
 
 describe('abacus-timewindow', () => {
   it('Calculate the index of a time window', () => {
@@ -216,5 +218,40 @@ describe('abacus-timewindow', () => {
       deep.equal([5, 4, 3, 2, 1]);
     expect(time.adjustWindows(timewindow, twCreation, adjustTo)[4]).to.
       deep.equal(['Last']);
+  });
+
+  context('when windows sizes are provided', () => {
+
+    it('expect windows to be initialized', () => {
+      const timeWindow = createTimeWindow({}, { [dimension.min]: 2 });
+      const windows = timeWindow.getWindows(dimension.min);
+      expect(windows).to.deep.equal([null, null]);
+    });
+
+  });
+
+  context('when windows sizes are not provided', () => {
+
+    context('when slack period is within its dimension', () => {
+
+      it('expect slack + 1 windows intitialized', () => {
+        const timeWindow = createTimeWindow(
+          { scale : dimension.min, width : 3 });
+        const windows = timeWindow.getWindows(dimension.min);
+        expect(windows).to.deep.equal([null, null, null, null]);
+      });
+
+    });
+
+    context('when slack period afects the next dimension', () => {
+
+      it('expect (width div <dimension size>) + 1 windows initialized', () => {
+        const timeWindow = createTimeWindow(
+          { scale : dimension.day, width : 30 });
+        const windows = timeWindow.getWindows(dimension.month);
+        expect(windows).to.deep.equal([null, null, null]);
+      });
+
+    });
   });
 });
